Rename io.github.ethersync => org.teamtype
