+++ .github/workflows/build.yaml
@@ -32,6 +32,6 @@ jobs:
            uses: softprops/action-gh-release@v2
            with:
               files: |
-                  ./build/distributions/ethersync-0.7.0-SNAPSHOT.zip
+                  ./build/distributions/teamtype-0.9.0-SNAPSHOT.zip
               tag_name: nightly
               prerelease: true
