--- src/main/kotlin/io/github/ethersync/Icons.kt
@@ -1,10 +1,0 @@
-package io.github.ethersync
-
-import com.intellij.openapi.util.IconLoader
-
-object Icons {
-
-   @JvmField
-   val ToolbarIcon = IconLoader.getIcon("/toolbar.svg", javaClass)
-
-}
