--- src/main/kotlin/io/github/ethersync/protocol/DocumentCloseRequest.kt
@@ -1,8 +1,0 @@
-package io.github.ethersync.protocol
-
-import com.google.gson.annotations.SerializedName
-
-data class DocumentCloseRequest(
-   @SerializedName("uri")
-   val documentUri: String,
-)
