--- src/main/kotlin/io/github/ethersync/protocol/DocumentOpenRequest.kt
@@ -1,9 +1,0 @@
-package io.github.ethersync.protocol
-
-import com.google.gson.annotations.SerializedName
-
-data class DocumentOpenRequest(
-   @SerializedName("uri")
-   val documentUri: String,
-   val content: String
-)
