--- src/main/kotlin/io/github/ethersync/protocol/EthersyncEditorProtocol.kt
@@ -1,12 +1,0 @@
-package io.github.ethersync.protocol
-
-import org.eclipse.lsp4j.jsonrpc.services.JsonNotification
-
-interface EthersyncEditorProtocol {
-
-   @JsonNotification
-   fun cursor(cursorEvent: CursorEvent)
-
-   @JsonNotification
-   fun edit(editEvent: EditEvent)
-}
