--- src/main/kotlin/io/github/ethersync/protocol/RemoteEthersyncClientProtocol.kt
@@ -1,20 +1,0 @@
-package io.github.ethersync.protocol
-
-import com.google.gson.JsonObject
-import org.eclipse.lsp4j.jsonrpc.services.JsonNotification
-import org.eclipse.lsp4j.jsonrpc.services.JsonRequest
-import java.util.concurrent.CompletableFuture
-
-interface RemoteEthersyncClientProtocol {
-    @JsonRequest
-    fun cursor(cursorRequest: CursorRequest): CompletableFuture<JsonObject>
-
-    @JsonRequest
-    fun open(documentRequest: DocumentOpenRequest): CompletableFuture<JsonObject>
-
-    @JsonRequest
-    fun edit(editRequest: EditRequest): CompletableFuture<JsonObject>
-
-    @JsonNotification
-    fun close(documentRequest: DocumentCloseRequest)
-}
