+++ src/main/kotlin/org/teamtype/EthersyncService.kt
@@ -1,0 +1,12 @@
+package org.teamtype
+
+import kotlinx.coroutines.Job
+
+interface EthersyncService {
+
+   fun start(joinCode: String?): Job
+
+   fun shutdown(): Job
+
+   fun startWithCustomCommandLine(commandLine: String)
+}
