+++ src/main/kotlin/org/teamtype/Icons.kt
@@ -1,0 +1,10 @@
+package org.teamtype
+
+import com.intellij.openapi.util.IconLoader
+
+object Icons {
+
+   @JvmField
+   val ToolbarIcon = IconLoader.getIcon("/toolbar.svg", javaClass)
+
+}
