+++ src/main/kotlin/org/teamtype/StopEthersyncDaemonAction.kt
@@ -1,0 +1,15 @@
+package org.teamtype
+
+import com.intellij.icons.AllIcons
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.components.service
+
+class StopEthersyncDaemonAction : AnAction("Stpo daemon", "Stop running ethersync daemon", AllIcons.Run.Stop) {
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+
+        val service = project.service<EthersyncService>()
+        service.shutdown()
+    }
+}
