+++ src/main/kotlin/org/teamtype/protocol/CursorRequest.kt
@@ -1,0 +1,11 @@
+package org.teamtype.protocol
+
+import com.google.gson.annotations.SerializedName
+import org.eclipse.lsp4j.Range
+
+data class CursorRequest(
+   @SerializedName("uri")
+   val uri: String,
+   @SerializedName("ranges")
+   val ranges: List<Range>
+)
