+++ src/main/kotlin/org/teamtype/protocol/Delta.kt
@@ -1,0 +1,11 @@
+package org.teamtype.protocol
+
+import com.google.gson.annotations.SerializedName
+import org.eclipse.lsp4j.Range
+
+data class Delta(
+   @SerializedName("range")
+   val range: Range,
+   @SerializedName("replacement")
+   val replacement: String
+)
