+++ src/main/kotlin/org/teamtype/protocol/DocumentCloseRequest.kt
@@ -1,0 +1,8 @@
+package org.teamtype.protocol
+
+import com.google.gson.annotations.SerializedName
+
+data class DocumentCloseRequest(
+   @SerializedName("uri")
+   val documentUri: String,
+)
