+++ src/main/kotlin/org/teamtype/protocol/DocumentOpenRequest.kt
@@ -1,0 +1,9 @@
+package org.teamtype.protocol
+
+import com.google.gson.annotations.SerializedName
+
+data class DocumentOpenRequest(
+   @SerializedName("uri")
+   val documentUri: String,
+   val content: String
+)
