+++ src/main/kotlin/org/teamtype/protocol/EditEvent.kt
@@ -1,0 +1,12 @@
+package org.teamtype.protocol
+
+import com.google.gson.annotations.SerializedName
+
+data class EditEvent(
+   @SerializedName("uri")
+   val documentUri: String,
+   @SerializedName("revision")
+   val editorRevision: UInt,
+   @SerializedName("delta")
+   val delta: List<Delta>,
+)
