+++ src/main/kotlin/org/teamtype/protocol/EditRequest.kt
@@ -1,0 +1,12 @@
+package org.teamtype.protocol
+
+import com.google.gson.annotations.SerializedName
+
+data class EditRequest(
+   @SerializedName("uri")
+   val documentUri: String,
+   @SerializedName("revision")
+   val revision: UInt,
+   @SerializedName("delta")
+   val delta: List<Delta>,
+)
