+++ src/main/kotlin/org/teamtype/protocol/EthersyncEditorProtocol.kt
@@ -1,0 +1,12 @@
+package org.teamtype.protocol
+
+import org.eclipse.lsp4j.jsonrpc.services.JsonNotification
+
+interface EthersyncEditorProtocol {
+
+   @JsonNotification
+   fun cursor(cursorEvent: CursorEvent)
+
+   @JsonNotification
+   fun edit(editEvent: EditEvent)
+}
